pyproject.toml
[tool.poetry]
name = "dogfighter"
version = "0.1.0"
description = ""
authors = ["Jet <taijunjet@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
torch = "^2.3.1"
jj-wingman = "^0.19.3"
gymnasium = "^1.0.0a2"
pydantic = "^2.7.4"
numpy = "<2.0.0"
tqdm = "^4.66.4"
imageio = {extras = ["ffmpeg"], version = "^2.34.2"}
pettingzoo = "^1.24.3"
shimmy = "^2.0.0"
dm-control = "^1.0.22"
prefetch-generator = "^1.0.3"

[tool.poetry.dev-dependencies]
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
